Use shelve-based cache for kafka. move common methods to base class
